ok now all the pages are seperated, and sidebar stuff commented out
